Fix meta panel not being movable
This was a two fold problem.

Firstly, when the movable panel was moved from the edit button to the meta panel the meta panel was left on the wrong layer, so it would never collide with the interaction point.

Secondly, the MovablePanel never got trigger collider calls from the colliders in the nested buttons. Presumably because the button handlers on those nested objects are using them. The movable panel needed it's own box collider to be grabbed.

Fixes #42