Fix a bug in meta panel rendering
- Meta panel camera was enabled
- Meta panel on-demand renderer was targeting spawn panel backdrop
- OnDemand renderers are invisible if there are no changes after enable to trigger render

As a result. The meta panel was being continually rendered by the camera instead of on-demand. And every time the on-demand renderer triggered a change it re-rendered the spawn panel instead.

Because of the OnDemand renderer issue the meta panel would have been invisible if it was actually being rendered on-demand.

To fix this:

- OnDemand renderers now consider the renderee dirty on the renderers OnEnable to ensure a first render is always done
- Meta panel camera is disabled
- Meta panel on-demand renderer target is fixed