Implement a state-based approach to throttle deadzone
The direction the throttle first leaves the deadzone from determines its
initial state, IE: Forward or Reverse. In this state, it is easy to fall
back into an associated "idle detent" state by moving the throttle
within the deadzone, eg: ForwardIdleDetent or ReverseIdleDetent.

From the "idle detent" state, it is easy to move into the associated
acceleration state, but "hard" to move to the opposite acceleration
state.

Transition between states can be made clear to the user by use of haptic
events.

This mechanism enables you to quickly move from forward to reverse if
you need to, while also making it easy to find (and not move past!)
idle.