merged Holographic Button into Holographic Overlay
Continuing merging files that were almost identical. Holographic Overlay can now be used for every instance of free-standing Textures that need to be rendered and is implemented on the ControlButton prefab.

I also removed the backface and added a default texture to the ControlButtonAssetCatalog. That way control buttons will receive a default texture if there isn't one defined at runtime, which removes the need to assign a backface to the Instantiated ControlButton prefabs.