new OverlayManager
separated out the overlay management into:
- static Locations (panels, joysticks, etc.)
- controlButtons
- settings (not complete)

Each component of the overlay manager now works in isolation to the others and has been converted to a prefab. The prefabs still live underneath the OverlayManager, but this makes testing and adding further capabilities much simpler. Overlay Management has also been removed from the SeatedOrigin.

For both the Static Locations and ControlButtons, there are anchor components that are used to help the corresponding manager to detect them. This decouples the classes from each other and allows for more flexible additions of tracked objects in the future.

SavedState still reads from the same file as before and the "key" element of the staticLcoations is used to connect the old hard-coded names to the new anchor method.

Settings are not finished yet. As it stands, they were not actually utilized, so I stripped them out for the moment. I intend to build a new system for user configurable settings that relies on the DesktopUI instead.