Add controlButton through DesktopUI
you can now use the desktopUI to add controlbuttons. This uses the gameevent system, so that both the savedstate file and the overlay can be updated independently.

The spawn manager uses the savedState file to find the next available location and places an entry into the file directly. Components that listen for the ControlButtonAdded event then use those same coordinates to add the controlButton to the scene or to the Desktop UI.

Also removed the SettingsView because those references weren't working