Changed GuiFocus default to NoFocus
Was previously Unknown, which isn't actually a value in the Status.json file.

The Status file will occasionally omit the GuiFocus key and we'd rather have the Overlay assume NoFocus than Unknown because that's what the game assumes in that same situation. The Status File briefly populates the GuiFocus value with 0 prior to updating it to the current value when the game launches, so our overlay will match this behavior.